feat: update score calcualtion
