+++ server/src/routes/course/score/index.ts
@@ -1,0 +1,6 @@
+export * from './createMultipleScores';
+export * from './createSingleScore';
+export * from './getScore';
+export * from './getScoreAsCsv';
+export * from './getScoreByStudent';
+export * from './recalculateScore';
